[bd-k4w] Fix 4 E2E failures: SUM BIGINT cast, reversal UUID shadow, oracle ar_invoice_id, trial balance optional currency

+++ e2e-tests/tests/common/mod.rs
@@ -317,8 +317,8 @@ pub async fn assert_journal_balanced(
 ) -> Result<(), String> {
     let row = sqlx::query(
         "SELECT
-            COALESCE(SUM(debit_minor), 0) as total_debits,
-            COALESCE(SUM(credit_minor), 0) as total_credits
+            COALESCE(SUM(debit_minor), 0)::BIGINT as total_debits,
+            COALESCE(SUM(credit_minor), 0)::BIGINT as total_credits
          FROM journal_lines
          WHERE journal_entry_id = $1"
     )
